Renamed class. Renamed executable and header file name (primary content is now class of same name). Some style adjustments based on line character limits. Revised code so that GSAP initialization is only performed once. Removed unneccasry for loop. Code compiles but causes power system node crash in current state.
